ProgressCube
Oh my goodness, nice to finally make some progress :') Isn't drawing correctly but I now have a cube. Assuming it's something I've done wrong declaring vertices or some shit coz I can see the triangles I need going to the wrong side -_- tbc...

+++ GameApp/src/TutorialScene.h
@@ -45,6 +45,6 @@ private:
 	float handRotationAngle = 0;
 	float handScale = 1;
 
-	int discSegments = 3;
+	float discSegments = 3.f;
 	float radius = 2;
 };
